feat: #526 remove legacy imports configuration
Move the only used function to liferay-npm-imports-checker.

Note that liferay-npm-imports-checker needs to be rewritten or removed for
bundler 3.
